+++ src/plugins/external-search/javtrailers.js
@@ -1,60 +1,3 @@
-const ie = class e {
-    constructor() {
-        if (new.target === e) throw new Error("HotkeyManager cannot be instantiated.");
-    }
-    static registerHotkey(e, t, n = null) {
-        if (Array.isArray(e)) {
-            let a = [];
-            return e.forEach((e => {
-                if (!this.isHotkeyFormat(e)) throw new Error("快捷键格式错误");
-                let i = this.recordHotkey(e, t, n);
-                a.push(i);
-            })), a;
-        }
-        if (!this.isHotkeyFormat(e)) throw new Error("快捷键格式错误");
-        return this.recordHotkey(e, t, n);
-    }
-    static recordHotkey(e, t, n) {
-        let a = Math.random().toString(36).substr(2);
-        return this.registerHotKeyMap.set(a, {
-            hotkeyString: e,
-            callback: t,
-            keyupCallback: n
-        }), a;
-    }
-    static unregisterHotkey(e) {
-        this.registerHotKeyMap.has(e) && this.registerHotKeyMap.delete(e);
-    }
-    static isHotkeyFormat(e) {
-        return e.toLowerCase().split("+").map((e => e.trim())).every((e => [ "ctrl", "shift", "alt" ].includes(e) || 1 === e.length));
-    }
-    static judgeHotkey(e, t) {
-        const n = e.toLowerCase().split("+").map((e => e.trim())), a = n.includes("ctrl"), i = n.includes("shift"), s = n.includes("alt"), o = n.find((e => "ctrl" !== e && "shift" !== e && "alt" !== e));
-        return (this.isMac ? t.metaKey : t.ctrlKey) === a && t.shiftKey === i && t.altKey === s && t.key.toLowerCase() === o;
-    }
-};
-
-i(ie, "isMac", 0 === navigator.platform.indexOf("Mac")), i(ie, "registerHotKeyMap", new Map),
-i(ie, "handleKeydown", (e => {
-    for (const [t, n] of ie.registerHotKeyMap) {
-        let t = n.hotkeyString, a = n.callback;
-        ie.judgeHotkey(t, e) && a(e);
-    }
-})), i(ie, "handleKeyup", (e => {
-    for (const [t, n] of ie.registerHotKeyMap) {
-        let t = n.hotkeyString, a = n.keyupCallback;
-        a && (ie.judgeHotkey(t, e) && a(e));
-    }
-}));
-
-let se = ie;
-
-document.addEventListener("keydown", (e => {
-    se.handleKeydown(e);
-})), document.addEventListener("keyup", (e => {
-    se.handleKeyup(e);
-}));
-
 class oe extends X {
     getName() {
         return "JavTrailersPlugin";
@@ -80,12 +23,6 @@ class oe extends X {
         })), window.addEventListener("message", (e => {
             let t = document.getElementById("vjs_video_3_html5_api");
             t && (t.currentTime += 5);
-        }));
-        const n = new URLSearchParams(window.location.search), a = n.get("filterHotKey"), i = n.get("favoriteHotKey"), s = n.get("speedVideoHotKey");
-        a && se.registerHotkey(a, (() => window.parent.postMessage(a, "*"))), i && se.registerHotkey(i, (() => window.parent.postMessage(i, "*"))),
-        s && se.registerHotkey(s, (() => {
-            const e = document.getElementById("vjs_video_3_html5_api");
-            e && (e.currentTime += 5);
         }));
     }
     handlePlayJavTrailers() {